api_yamdb/requirements.txt
asgiref==3.2.10
requests
django==2.2.16
django_extensions
django-filter==2.4.0
djangorestframework-simplejwt==4.8.0
djangorestframework==3.11.2
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==2.1.0
pytz==2020.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
sqlparse==0.3.1
python-dotenv==0.19.1
djoser==2.1.0
flake8
pep8-naming
flake8-broken-line
flake8-return
flake8-isort